requirements.txt
openai==0.27.2
python-decouple==3.8
requests==2.28.2
tqdm==4.65.0
